add multi-facilities contact list
