+++ src/applications/mhv-secure-messaging/tests/e2e/fixtures/all-recipients-response.json
@@ -2,7 +2,7 @@
     "data": [
       {
         "id": "7026562",
-        "type": "al_triage_teams",
+        "type": "all_triage_teams",
         "attributes": {
           "triageTeamId": 7026562,
           "name": "###ABC_XYZ_TRIAGE_TEAM_PCMM_ASSOCIATION_747###",
@@ -14,7 +14,7 @@
       },
       {
         "id": "6832726",
-        "type": "al_triage_teams",
+        "type": "all_triage_teams",
         "attributes": {
           "triageTeamId": 6832726,
           "name": "###PQR TRIAGE_TEAM 747###",
@@ -26,7 +26,7 @@
       },
       {
         "id": "6910405",
-        "type": "al_triage_teams",
+        "type": "all_triage_teams",
         "attributes": {
           "triageTeamId": 6910405,
           "name": "***Jeasmitha-Cardio-Clinic***",
@@ -38,7 +38,7 @@
       },
       {
         "id": "6978853",
-        "type": "al_triage_teams",
+        "type": "all_triage_teams",
         "attributes": {
           "triageTeamId": 6978853,
           "name": "***Mohammad_Fort Belovoir TG 100_SLC4%",
@@ -50,7 +50,7 @@
       },
       {
         "id": "6978854",
-        "type": "al_triage_teams",
+        "type": "all_triage_teams",
         "attributes": {
           "triageTeamId": 6978854,
           "name": "***Mohammad_Fort Belovoir TG 200_APPT_SLC4%",
@@ -62,7 +62,7 @@
       },
       {
         "id": "7123233",
-        "type": "al_triage_teams",
+        "type": "all_triage_teams",
         "attributes": {
           "triageTeamId": 7123233,
           "name": "CAMRY_PCMM RELATIONSHIP_05092022_SLC4",
@@ -74,7 +74,7 @@
       },
       {
         "id": "56219",
-        "type": "al_triage_teams",
+        "type": "all_triage_teams",
         "attributes": {
           "triageTeamId": 56219,
           "name": "Liana's SLC4 PCMM",
