+++ src/applications/mhv-secure-messaging/tests/e2e/fixtures/multi-facilities-recipients-response.json
@@ -6,7 +6,7 @@
       "attributes": {
         "triageTeamId": 7026562,
         "name": "###ABC_XYZ_TRIAGE_TEAM_PCMM_ASSOCIATION_747###",
-        "stationNumber": "989",
+        "stationNumber": "531",
         "blockedStatus": false,
         "relationType": "PATIENT",
         "preferredTeam": true
@@ -18,7 +18,7 @@
       "attributes": {
         "triageTeamId": 6832726,
         "name": "TG-7410",
-        "stationNumber": "989",
+        "stationNumber": "531",
         "blockedStatus": false,
         "relationType": "PATIENT",
         "preferredTeam": true
@@ -30,7 +30,7 @@
       "attributes": {
         "triageTeamId": 6910405,
         "name": "Jeasmitha-Cardio-Clinic",
-        "stationNumber": "989",
+        "stationNumber": "531",
         "blockedStatus": false,
         "relationType": "PATIENT",
         "preferredTeam": true
@@ -41,7 +41,7 @@
       "type": "all_triage_teams",
       "attributes": {
         "triageTeamId": 6978853,
-        "stationNumber": "989",
+        "stationNumber": "583",
         "blockedStatus": false,
         "name": "***TG 100_SLC4%",
         "relationType": "PATIENT",
@@ -53,7 +53,7 @@
       "type": "all_triage_teams",
       "attributes": {
         "triageTeamId": 6978854,
-        "stationNumber": "989",
+        "stationNumber": "583",
         "blockedStatus": false,
         "name": "***TG 200_APPT_SLC4%",
         "relationType": "PATIENT",
@@ -66,7 +66,7 @@
       "attributes": {
         "triageTeamId": 7123233,
         "name": "CAMRY_PCMM RELATIONSHIP_05092022_SLC4",
-        "stationNumber": "989",
+        "stationNumber": "583",
         "blockedStatus": false,
         "relationType": "PATIENT",
         "preferredTeam": true
@@ -77,7 +77,7 @@
       "type": "all_triage_teams",
       "attributes": {
         "triageTeamId": 56219,
-        "stationNumber": "989",
+        "stationNumber": "583",
         "blockedStatus": false,
         "name": "SLC4 PCMM",
         "relationType": "PATIENT",
@@ -90,7 +90,7 @@
       "attributes": {
         "triageTeamId": 3658288,
         "name": "Record Amendment Admin",
-        "stationNumber": "989",
+        "stationNumber": "531",
         "blockedStatus": false,
         "preferredTeam": true,
         "relationshipType": "PATIENT"
@@ -102,7 +102,7 @@
       "attributes": {
         "triageTeamId": 3658299,
         "name": "OH TG GROUP 002",
-        "stationNumber": "989",
+        "stationNumber": "531",
         "blockedStatus": false,
         "preferredTeam": true,
         "relationshipType": "PATIENT"
