+++ src/applications/mhv-secure-messaging/tests/e2e/secure-messaging-contact-list-interface.cypress.spec.js
@@ -2,15 +2,17 @@ import SecureMessagingSite from './sm_site/SecureMessagingSite';
 import PatientInboxPage from './pages/PatientInboxPage';
 import ContactListPage from './pages/ContactListPage';
 import { AXE_CONTEXT } from './utils/constants';
+import mockMixedCernerFacilitiesUser from './fixtures/userResponse/user-cerner-mixed.json';
+import mockFacilities from './fixtures/facilityResponse/cerner-facility-mock-data.json';
+import mockEhrData from './fixtures/userResponse/vamc-ehr-cerner-mixed.json';
+import mockMixRecipients from './fixtures/multi-facilities-recipients-response.json';
 
 describe('Secure Messaging Contact list', () => {
-  beforeEach(() => {
+  it('verify base web-elements - single facility', () => {
     SecureMessagingSite.login();
     PatientInboxPage.loadInboxMessages();
     ContactListPage.loadContactList();
-  });
 
-  it('verify base web-elements', () => {
     ContactListPage.verifyHeaders();
     ContactListPage.verifyAllCheckboxes(true);
     ContactListPage.selectAllCheckBox();
@@ -21,73 +23,22 @@ describe('Secure Messaging Contact list', () => {
     cy.axeCheck(AXE_CONTEXT);
   });
 
-  it('verify contact list alerts', () => {
+  it(`verify base web-elements - multi facilities`, () => {
+    SecureMessagingSite.login(
+      mockEhrData,
+      true,
+      mockMixedCernerFacilitiesUser,
+      mockFacilities,
+    );
+    PatientInboxPage.loadInboxMessages();
+    ContactListPage.loadContactList(mockMixRecipients);
+
+    ContactListPage.verifyHeaders();
+    ContactListPage.verifyAllCheckboxes(true);
     ContactListPage.selectAllCheckBox();
-    ContactListPage.clickCancelButton();
-    ContactListPage.verifySaveAlertHeader();
+    ContactListPage.verifyAllCheckboxes(false);
     ContactListPage.verifyButtons();
-    ContactListPage.closeSaveModal();
-
-    ContactListPage.clickBackToInbox();
-    ContactListPage.verifySaveAlertHeader();
-    ContactListPage.verifyButtons();
-    ContactListPage.closeSaveModal();
-
-    ContactListPage.clickSaveAndExitButton();
-    ContactListPage.verifyEmptyContactListAlert();
-
     cy.injectAxe();
     cy.axeCheck(AXE_CONTEXT);
-  });
-
-  it('verify single contact selected', () => {
-    ContactListPage.selectAllCheckBox();
-    ContactListPage.selectCheckBox(`100`);
-
-    ContactListPage.clickSaveAndExitButton();
-    ContactListPage.verifyContactListSavedAlert();
-
-    cy.wait('@savedList')
-      .its('request.body')
-      .then(req => {
-        const selected = req.updatedTriageTeams.filter(el =>
-          el.name.includes(`100`),
-        );
-
-        cy.wrap(selected).each(el => {
-          expect(el.preferredTeam).to.eq(true);
-        });
-      });
-
-    cy.injectAxe();
-    cy.axeCheck(AXE_CONTEXT);
-  });
-
-  it(`verify few contacts selected`, () => {
-    ContactListPage.selectAllCheckBox();
-    ContactListPage.selectCheckBox(`100`);
-    ContactListPage.selectCheckBox(`Cardio`);
-    ContactListPage.selectCheckBox(`TG-7410`);
-    ContactListPage.clickSaveAndExitButton();
-
-    ContactListPage.verifyContactListSavedAlert();
-
-    cy.wait('@savedList')
-      .its('request.body')
-      .then(req => {
-        const selected = req.updatedTriageTeams.filter(
-          el =>
-            el.name.includes(`100`) ||
-            el.name.includes(`Cardio`) ||
-            el.name.includes(`TG-7410`),
-        );
-
-        cy.wrap(selected).each(el => {
-          expect(el.preferredTeam).to.eq(true);
-        });
-
-        cy.injectAxe();
-        cy.axeCheck(AXE_CONTEXT);
-      });
   });
 });
